+++ src/Compute/Compute.Test/ScenarioTests/VirtualMachineTests.cs
@@ -675,5 +675,12 @@ namespace Microsoft.Azure.Commands.Compute.Test.ScenarioTests
         {
             TestRunner.RunTestScript("Test-VMSetAzOSCredentialNullRef");
         }
-    }
+
+        [Fact]
+        [Trait(Category.AcceptanceType, Category.CheckIn)]
+        public void testgensetazvm()
+        {
+            TestRunner.RunTestScript("TestGen-setazvm");
+        }
+        }
 }